feat ✨: Update dependencies and add new Services component.
- Added new "Services" component to the home page.
- Added new Services component with a single service information.
- Updated dependencies to their latest versions.
- It looks like you've provided a snippet of a package-lock.json file, which is used by npm to keep track of the dependencies required by your project. Here's a breakdown of what I can see:

**Package updates**

The first part of the code suggests that there are some changes to the dependencies being used in the project.

* The `react-icons` package has been updated from version 5.2.0 to 5.3.0.
* The `tailwindcss` package has been updated from version 3.4.1 to 3.4.13 (a minor update).
* The `vaul` package has been updated from version 0.9.4 to 0.9.9.

**Checksums and hashes**

The code includes checksums for each of the updated packages, which are used by npm to verify the integrity of the dependencies being installed. These checksums are hexadecimal strings that represent the hash of the package's contents.

**Language and link type**

For each package, there is a `languageName` field set to "node", indicating that these packages are JavaScript modules intended for use in Node.js environments. The `linkType` field is also set to "hard", which means that npm will install these dependencies as hard links on the file system.

**Peer dependencies**

For some packages, there are peer dependencies specified, which are other packages that must be installed alongside them. For example, the `vaul` package requires `react` and `react-dom` to be installed.

Overall, this code snippet appears to be a standard part of the package-lock.json file used by npm to manage project dependencies.